backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.17.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
brotli==1.1.0
tavily-python==0.3.0

# Reddit API
praw==7.7.1
asyncpraw==7.7.1

# Data processing and analysis
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Date and time handling
pytz==2023.3
python-dateutil==2.8.2

# HTTP client for API calls
aiohttp==3.9.1
requests>=2.32.3

# Google API integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0
gspread-dataframe==3.3.1

# Database support - Prioritize async drivers
# asyncpg==0.29.0  # Already listed above
# psycopg2-binary==2.9.9  # Commented out to avoid conflicts
mysql-connector-python==8.2.0

# Email support
aiosmtplib==3.0.1

# Image processing
Pillow==10.1.0

# Text analysis and NLP
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9

# Translation (using deep-translator instead of googletrans for compatibility)
deep-translator==1.11.4

# Web scraping
beautifulsoup4==4.12.2

# Enhanced web crawling and content extraction
playwright==1.40.0
easyocr==1.7.0
pytesseract==0.3.10

# Speech-to-text functionality
openai-whisper==20250625

# High-performance content extraction
extractous==0.3.0

# PDF processing
PyPDF2==3.0.1
reportlab==4.0.7
PyMuPDF==1.23.8

# Social media APIs
tweepy==4.14.0
facebook-business==18.0.0

# Payment processing
stripe==7.8.0

# Notifications
twilio==8.10.0
firebase-admin==6.2.0

# Template engine
jinja2==3.1.2

# OpenAI integration
openai==1.3.7

# Vector database and document processing
chromadb==0.4.18
sentence-transformers==2.2.2

# MongoDB support
pymongo==4.6.0

# Document processing
python-docx==1.1.0

# PDF generation
markdown==3.5.1

# RSS feed parsing
feedparser==6.0.10

# Vercel blob storage
vercel-blob==0.4.2 
telethon==1.34.0


# YouTube transcript extraction
youtube-transcript-api==0.6.2
